+++ packages/cursorless-engine/src/core/HatTokenMapImpl.ts
@@ -155,24 +155,3 @@ export class HatTokenMapImpl implements HatTokenMap {
     this.prePhraseMapsSnapshotTimestamp = hrtime.bigint();
   }
 }
-
-export class DisabledHatTokenMap implements HatTokenMap {
-  async allocateHats() {
-    // Do nothing
-  }
-
-  async getReadableMap() {
-    return {
-      getEntries() {
-        return [];
-      },
-      getToken() {
-        throw new Error("Hat map is disabled");
-      },
-    };
-  }
-
-  dispose() {
-    // Do nothing
-  }
-}
